+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+    - [meta] added possibility for `priority` to be empty (GitHub Issue #76)
+    - [config] added `priority_default` field (GitHub Issue #76)
     - [cmd/list] added possibility to filter for ranges: `>` and `<` (GitHub Issue #75)
     - [cmd/list] added support for relationships (GitHub Issue #59)
     - [config] added possibility to change ID generation strategy `id_generation` (GitHub Issue #9)